# semconv: mark the package as deprecated