--- a/semconv/v1.22.0/doc.go
+++ b/semconv/v1.22.0/doc.go
@@ -6,4 +6,6 @@
 // OpenTelemetry semantic conventions are agreed standardized naming
 // patterns for OpenTelemetry things. This package represents the v1.22.0
 // version of the OpenTelemetry semantic conventions.
+//
+// Deprecated: [v0.126.0] This package has been deprecated, use go.opentelemetry.io/otel/semconv instead.
 package semconv // import "go.opentelemetry.io/collector/semconv/v1.22.0"