This package will be removed in the future, use go.opentelemetry.io/otel/semconv instead