@codeboten 2025-05-11
@bogdandrutu sure, i created the following PRs to update the core/contrib repos to use the otelgo semconv package:
* https://github.com/open-telemetry/opentelemetry-collector/pull/12991
* https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/39979
* https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/39986

It seemed to work ok. See the comment in the original PR:
> This updates the calls to the collector's internal semconv package with otelgo's instead. The main difference is how the keys can be used, they will need to be cast as strings where they key name is needed, otherwise it's not a huge change. This would allow us to stop producing our own semconv package once contrib is moved as well.

I ran some benchmarks of instantiating the service's Resource using the old semconv and the new ones to validate that using the new semconv wouldn't have an impact on allocations:

```
benchstat 20_runs_bench.txt 20_oldsemconv_runs_bench.txt 
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/service/internal/resource
cpu: Apple M3 Max
               │ 20_runs_bench.txt │    20_oldsemconv_runs_bench.txt    │
               │      sec/op       │   sec/op     vs base               │
NewResource-14         332.2n ± 2%   323.4n ± 1%  -2.68% (p=0.000 n=20)

               │ 20_runs_bench.txt │  20_oldsemconv_runs_bench.txt  │
               │       B/op        │    B/op     vs base            │
NewResource-14          720.0 ± 0%   720.0 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal

               │ 20_runs_bench.txt │  20_oldsemconv_runs_bench.txt  │
               │     allocs/op     │ allocs/op   vs base            │
NewResource-14          8.000 ± 0%   8.000 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal
```

Is there a specific aspect of pdata using the otel go's semconv that you're thinking about?