@bogdandrutu 2025-05-12
@codeboten 

Are we concerned about `string(semconv.ServiceNameKey)` being a bit too verbose/cryptic? Should we, if this is a pattern we want, at least extend the pcommon.Map to have some overloads that accepts as key the otel attrs? Or at least something like "PutKeyValue" for example?