@codeboten 2025-05-12
> Is this a release blocker ?

It doesn't have to block the release i don't think. We don't need to mark the semconv package as deprecated in this release